fix codex warning

+++ ecmascript/js_api_tree_set_iterator.cpp
@@ -53,7 +53,7 @@ JSTaggedValue JSAPITreeSetIterator::Next(EcmaRuntimeCallInfo *argv)
     }
 
     // Let index be Set.[[NextIndex]].
-    int index = iter->GetNextIndex();
+    int index = static_cast<int>(iter->GetNextIndex());
     if (index < elements) {
         IterationKind itemKind = IterationKind(iter->GetIterationKind());
